Добавил редактор сообщений, сломал шифрование паролей.

+++ src/main/resources/db/migration/V1__Init_DB.sql
@@ -1,7 +1,7 @@
 create sequence hibernate_sequence start 1 increment 1;
 
 create table message (
-    id int8 not null,
+    id int4 not null,
     filename varchar(255),
     tag varchar(255),
     text varchar(2048) not null,
